requirements.txt
streamlit
langchain
langchain_core
langchain_community
pillow
openai